Revert "Factor repair from gossip (#8044)"
This reverts commit e61257695fef39408061213d3b853b9b537b73b6.
