+++ core/src/lib.rs
@@ -43,8 +43,6 @@ pub mod rpc_pubsub_service;
 pub mod rpc_service;
 pub mod rpc_subscriptions;
 pub mod sendmmsg;
-pub mod serve_repair;
-pub mod serve_repair_service;
 pub mod sigverify;
 pub mod sigverify_shreds;
 pub mod sigverify_stage;
