+++ validator/src/main.rs
@@ -213,6 +213,7 @@ fn get_rpc_addr(
     let gossip_service = GossipService::new(
         &cluster_info.clone(),
         None,
+        None,
         node.sockets.gossip.try_clone().unwrap(),
         &gossip_exit_flag,
     );
